:art: Use sb-stream-promise and re-arrange the code
